Move util to `testutil.option.util`
