+++ src/python/pants/backend/awslambda/python/BUILD
@@ -32,7 +32,7 @@ python_tests(
     'src/python/pants/source',
     'src/python/pants/util:contextutil',
     'src/python/pants/testutil:test_base',
-    'src/python/pants/testutil/engine:util',
+    'src/python/pants/testutil/option',
   ],
   tags = {'partially_type_checked'},
 )
