+++ src/python/pants/backend/python/lint/black/BUILD
@@ -31,7 +31,7 @@ python_tests(
     'src/python/pants/rules/core',
     'src/python/pants/source',
     'src/python/pants/testutil:test_base',
-    'src/python/pants/testutil/engine:util',
+    'src/python/pants/testutil/option',
   ],
   tags = {'integration', 'partially_type_checked'},
   timeout = 120,
