+++ src/python/pants/backend/python/lint/flake8/rules_integration_test.py
@@ -16,8 +16,8 @@ from pants.engine.rules import RootRule
 from pants.engine.selectors import Params
 from pants.rules.core.lint import LintResult
 from pants.source.wrapped_globs import EagerFilesetWithSpec
-from pants.testutil.engine.util import create_options_bootstrapper
 from pants.testutil.interpreter_selection_utils import skip_unless_python27_and_python3_present
+from pants.testutil.option.util import create_options_bootstrapper
 from pants.testutil.test_base import TestBase
