+++ src/python/pants/backend/python/rules/pex_test.py
@@ -20,7 +20,7 @@ from pants.engine.fs import Digest, DirectoryToMaterialize, FileContent, InputFi
 from pants.engine.isolated_process import ExecuteProcessRequest, ExecuteProcessResult
 from pants.engine.rules import RootRule
 from pants.engine.selectors import Params
-from pants.testutil.engine.util import create_options_bootstrapper
+from pants.testutil.option.util import create_options_bootstrapper
 from pants.testutil.subsystem.util import init_subsystem
 from pants.testutil.test_base import TestBase
 from pants.util.strutil import create_path_env_var
