+++ src/python/pants/rules/core/BUILD
@@ -38,6 +38,7 @@ python_tests(
     'src/python/pants/testutil:test_base',
     'src/python/pants/testutil:goal_rule_test_base',
     'src/python/pants/testutil/engine:util',
+    'src/python/pants/testutil/option',
     'src/python/pants/testutil/subsystem',
   ],
   tags = {"partially_type_checked"},
